Additions to Consensus Seq
I thought it would be better to outline the specific steps for counting one of the nucleotides ("A") and see how they apply it to the other nucleotides, giving them the option of either adding to the current function they have or creating three more functions. I think it's probably best to emphasize the method like Neil and Claudia were saying, just so we don't have any completely whacky solutions!